Refactor: CommonException 위치 변경
- common 패키지 하위에 exception 패키지 추가
- CommonException을 common.exception 패키지로 이동
