+++ src/main/java/com/uranus/taskmanager/api/authentication/exception/AuthenticationException.java
@@ -2,7 +2,7 @@ package com.uranus.taskmanager.api.authentication.exception;
 
 import org.springframework.http.HttpStatus;
 
-import com.uranus.taskmanager.api.common.CommonException;
+import com.uranus.taskmanager.api.common.exception.CommonException;
 
 public abstract class AuthenticationException extends CommonException {
