+++ src/main/java/com/uranus/taskmanager/api/workspace/exception/WorkspaceException.java
@@ -2,7 +2,7 @@ package com.uranus.taskmanager.api.workspace.exception;
 
 import org.springframework.http.HttpStatus;
 
-import com.uranus.taskmanager.api.common.CommonException;
+import com.uranus.taskmanager.api.common.exception.CommonException;
 
 public abstract class WorkspaceException extends CommonException {
 	public WorkspaceException(String message, HttpStatus httpStatus) {
