+++ src/main/java/com/uranus/taskmanager/api/workspacemember/authorization/exception/AuthorizationException.java
@@ -2,7 +2,7 @@ package com.uranus.taskmanager.api.workspacemember.authorization.exception;
 
 import org.springframework.http.HttpStatus;
 
-import com.uranus.taskmanager.api.common.CommonException;
+import com.uranus.taskmanager.api.common.exception.CommonException;
 
 public abstract class AuthorizationException extends CommonException {
