+++ src/main/java/com/uranus/taskmanager/api/workspacemember/exception/WorkspaceMemberException.java
@@ -2,7 +2,7 @@ package com.uranus.taskmanager.api.workspacemember.exception;
 
 import org.springframework.http.HttpStatus;
 
-import com.uranus.taskmanager.api.common.CommonException;
+import com.uranus.taskmanager.api.common.exception.CommonException;
 
 public abstract class WorkspaceMemberException extends CommonException {
